TCES-12 refactor(auth) refactor all routes to use controller
