+++ server/src/configs/sequelize.js
@@ -23,6 +23,12 @@ const sequelize = new Sequelize(
   },
 );
 
+try {
+    sequelize.authenticate().then(() => { console.log('Connection has been established successfully.'); });
+} catch (error) {
+    console.error('Unable to connect to the database:', error);
+} 
+
 module.exports = {
     sequelize: sequelize
 }
